Resign all jar libary files so Toucan also works on linux when launched with IcedTea Java Web Start.
Resign all jar libary files so Toucan also works on linux when launched
with IcedTea Java Web Start, which seems to be a bit more picky than
Oracle/Sun Java Web Start.

Also the toucan jar file needs to be signed everytime it is recompiled.

The following commands and settings were used:

$ keytool -genkey -keystore keystore_toucan -alias toucan_key -validity 10000
Enter keystore password:
Re-enter new password:
What is your first and last name?
  [Unknown]:  Stein Aerts
What is the name of your organizational unit?
  [Unknown]:  University of Leuven
What is the name of your organization?
  [Unknown]:  University of Leuven
What is the name of your City or Locality?
  [Unknown]:  Leuven
What is the name of your State or Province?
  [Unknown]:  Vlaams-Brabant
What is the two-letter country code for this unit?
  [Unknown]:  BE
Is CN=Stein Aerts, OU=University of Leuven, O=University of Leuven, L=Leuven, ST=Vlaams-Brabant, C=BE correct?
  [no]:  no
What is your first and last name?
  [Stein Aerts]:
What is the name of your organizational unit?
  [University of Leuven]:  Lab of Computational Biology
What is the name of your organization?
  [University of Leuven]:
What is the name of your City or Locality?
  [Leuven]:
What is the name of your State or Province?
  [Vlaams-Brabant]:
What is the two-letter country code for this unit?
  [BE]:
Is CN=Stein Aerts, OU=Lab of Computational Biology, O=University of Leuven, L=Leuven, ST=Vlaams-Brabant, C=BE correct?
  [no]:  yes

Enter key password for <toucan_key>
	(RETURN if same as keystore password):

for jar in *.jar ; do
    jarsigner -keystore keystore_toucan ${jar} toucan_key;
done

for jar in *.jar ; do
    jarsigner -verify -certs -verbose ${jar};
done

Signed-off-by: Gert Hulselmans <gert.hulselmans@med.kuleuven.be>